Add -expose-package flag
The `-package` flag does two things:
* it makes the package avaliable, and
* it eagerly links the package when running under the interpreter.

When compiling a program using Template Haskell this is unnecessarily
expensive when only a small part of the packages is used for evaluating
TH splices.

This introduces a new flag that only makes the package avaliable and
since the interpreter already implements lazy loading everything still
works.

There are still reasons to want the behavior that `-package` provides, like
C library referring to a symbol from a Haskell package.

Test Plan:
contbuild
manual testing looking at 'Linking' lines in '-v' output for a TH program
I can write a test if anyone feels strongly about it.

Reviewers: simonmar, austin, ezyang

Subscribers: ezyang, thomie, bgamari

Differential Revision: https://phabricator.haskell.org/D973

GHC Trac Issues: #2437

(this is still not in GHC master)