pmCheck: Don't generate PmId OccNames from Uniques
Doug Wilson astutely noticed that the pattern match checker was spending
a significant amount of time in mkPmId. It turns out that it was
producing Ids with OccNames that were generated from a Unique, which are
non-trivial to render. Since Var is strict in Name, Name in OccName, and
OccName in its FastString all of this encoding work was being done
despite the fact that it was (as far as I can tell) never actually
needed.

Test Plan: Validate, note allocations of `T11195`

Reviewers: austin, gkaracha

Subscribers: rwbarton, thomie

Differential Revision: https://phabricator.haskell.org/D3568