Always retain bindings for interactive module
Summary: Retaining all bindings from interpreted modules consumes a lot of memory and we have seen this causing problems on some of our machines with less memory. This adds a flag to only retain modules for explicitly added targets. This set of modules is presumably much smaller than the full graph dependencies and should allow us to save a lot of memory.

Test Plan:
 - Built and test memory usage of ghci with `-fminimal-bindings-retention`

Reviewers: watashi, smarlow, #sigma

Reviewed By: smarlow

Subscribers: smarlow