[fbonly] recompile target modules on :add if bindings are missing
Summary:
Sometimes even when the obtained Linkable(s) from the previous
load are not stale, we still have to recompile the modules because
we don't have their full bindings available. This can happen with
the flag -fminimal-bindings-retention turned on as this flag caused
us to only retain bindings of modules in hsc_target field of HscEnv.

Test Plan:
```
./validate --build-only
./validate --testsuite-only
```

Reviewers: watashi