Fix #16392: revertCAFs in external interpreter when necessary
We revert CAFs when loading/adding modules in ghci (presumably to refresh
execution states and to allow for object code to be unloaded from the runtime).
However, with `-fexternal-interpreter` enabled, we are only doing it in the
ghci process instead of the external interpreter process where the cafs are
allocated and computed. This makes sure that revertCAFs is done in the
appropriate process no matter if that flag is present or not.

gitlab merge request https://gitlab.haskell.org/ghc/ghc/merge_requests/500
(cherry-pick form 7a68254a7284db5bf8f1fa82aba4a6825d8f050a)