read command-line options from @-file
**For future ghc maintainer, please take the solution from**

https://gitlab.haskell.org/ghc/ghc/issues/16476

**if that's done.**

This add support the ghc and ghc-pkg tools to accept cli argument `@file` and
reads the arguments from `file`, so that we won't hit the ARG_MAX limit.

The behavior here is compatible with other built tools like `gcc`, `javac`, `ld` and `ar`.

Note that this is not completely backward compatible, before you can have
`ghc -optP @flags` and `@flags` will be passed to and expanded by `cpp`, but now
it will be expanded by `ghc` and the cli arguments may become malformed.
The workaround is to use `ghc -optP=@flags`.