[fbonly] Force linker to evaluate new closure and itbl env on unload
On linker unload, due to the lazy nature the IntMap used by closure_env
and itbl_env, we can have dangling references to object code in the
iserv process. This cause those object code to be held onto and may
cause crashes as seen in #16525. In this diff, I force those to be
evaluated so that we can clean them up in ghci.