fix crash when objects are not being freed correctly
We found a crash in ghci while loading new files, the stack trace
points to info table of static objects being freed incorrectly.
This addresses the issue by adding an address check for the info table
as well in `CheckUnload.c`. This would make it so that we do not
incorrectly free unloaded ObjectCode struct that still contain live info
table.